salvando

+++ src/main/java/br/com/alura/domain/Agencia.java
@@ -16,8 +16,8 @@ public class Agencia {
     @JoinColumn(name = "endereco_id")
     private Endereco endereco;
 
-    public Integer getId() {
-        return id;
+    public Long getId() {
+        return Long.valueOf(id);
     }
 
     public String getNome() {
